feat: integrate create and read operations for Properties module

+++ server/tables/propertiesTable.js
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS properties (
     property_status VARCHAR(50) DEFAULT 'Available',
     base_rent DECIMAL(10,2),
     property_taxes_quarterly DECIMAL(10,2),
-    security_deposit_months DECIMAL(10,2),
+    security_deposit_months INT,
     minimum_lease_term_months INT DEFAULT 24,
     address_id INT,
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
